GH-503 extracted files
