+++ SafeUIKit/ReplaceBrowserExtensionUI/Common/FeeCalculation/FeeCalculation.swift
@@ -1,0 +1,25 @@
+//
+//  Copyright © 2019 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+public class FeeCalculation: ArrayBasedCollection<FeeCalculationSection> {
+
+    public func makeView() -> UIView {
+        let backgroundView = UIView()
+        backgroundView.translatesAutoresizingMaskIntoConstraints = false
+        addSections(to: backgroundView)
+        return backgroundView
+    }
+
+    private func addSections(to backgroundView: UIView) {
+        let stackView = UIStackView(arrangedSubviews: elements.map { $0.makeView() })
+        stackView.axis = .vertical
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        backgroundView.addSubview(stackView)
+        backgroundView.wrapAroundDynamicHeightView(stackView, insets: .zero)
+    }
+
+}
