+++ SafeUIKit/ReplaceBrowserExtensionUI/Common/FeeCalculation/FeeCalculationLine.swift
@@ -1,0 +1,17 @@
+//
+//  Copyright © 2019 Gnosis Ltd. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+public class FeeCalculationLine {
+
+    var lineHeight: Double = 25
+
+    func makeView() -> UIView {
+        return UIView()
+    }
+
+}
+
